third_party/blink/web_tests/fast/dom/documenturi-readonly-expected.txt

ALERT: PASS: document.documentURI is read only.
This test ensures that document.documentURI is a readonly attribute. It passes if a "PASS" alert is displayed.
